+++ CHANGES.rst
@@ -17,6 +17,7 @@ Unreleased
 * Added :class:`.CalendarConfiguration` to represent the configuration of a
   :class:`.Calendar` widget.
 * Added :class:`.Hover` to represent the hover state of a :class:`.Button`.
+* Added :class:`.Styles` to represent widget styling information.
 
 **Changed**